requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1.tar.gz
